Update 11029 - Leading and Trailing.cpp

+++ 11029 - Leading and Trailing.cpp
@@ -1,9 +1,9 @@
 #include<bits/stdc++.h>
 #define ll long long int
 using namespace std;
-ll binpow(ll a,ll b)
+int binpow(ll a,ll b)
 {
-    ll res = 1;
+    int res = 1;
     while(b>0)
     {
         if(b&1)
